third_party/WebKit/LayoutTests/shadow-dom/v0/get-destination-insertion-points-re-distribution-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests for getDestinationInsertionPoints() which involves re-distribution.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS nodeList.item(1) is expectedNodes[1]
PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS document.getElementById("child-12").getDestinationInsertionPoints().length is 0
PASS successfullyParsed is true

TEST COMPLETE
